Add controllers, separate routes
