+++ src/server.js
@@ -1,11 +1,10 @@
 const express = require('express');
 const appRouter = require('./routes')
-// const express = require('express');
-// const router = express.Router();
+
 const app = express();
 
 let port;
-
+app.use('/', appRouter);
 const server = app.listen(3000, () => {
 	port = server.address().port;
 	console.log('Server ready at port %s', port);
